pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Helena Oliveira", email = "helena@informach.com.br"},
]
dependencies = [
    "Django>=4.0.5",
    "djangorestframework>=3.13.1",
]
requires-python = ">=3.9"
license = {text = "MIT"}

[tool]
[tool.pdm]

[build-system]
requires = ["pdm-pep517>=0.12.0"]
build-backend = "pdm.pep517.api"
